unify more tests
