+++ remote/client_test.go
@@ -1,12 +1,15 @@
 package remote
 
 import (
+	"bytes"
 	"context"
+	"io/ioutil"
+	"os"
+	"path/filepath"
 	"runtime"
 	"strings"
 	"syscall"
 	"testing"
-	"time"
 
 	"github.com/mongodb/grip"
 	"github.com/mongodb/grip/level"
@@ -30,7 +33,7 @@ type ClientTestCase struct {
 	ShouldSkip bool
 }
 
-func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
+func AddBasicClientTests(modify testutil.OptsModify, tests ...ClientTestCase) []ClientTestCase {
 	return append([]ClientTestCase{
 		{
 			Name: "ValidateFixture",
@@ -49,6 +52,7 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 			Name: "ProcEnvVarMatchesManagerID",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
 				opts := testutil.TrueCreateOpts()
+				modify(opts)
 				proc, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
 				info := proc.Info(ctx)
@@ -67,7 +71,9 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 		{
 			Name: "CreateProcessFails",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
-				proc, err := client.CreateProcess(ctx, &options.Create{})
+				opts := &options.Create{}
+				modify(opts)
+				proc, err := client.CreateProcess(ctx, opts)
 				require.Error(t, err)
 				assert.Nil(t, proc)
 			},
@@ -76,7 +82,9 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 			Name: "ListAllReturnsErrorWithCanceledContext",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
 				cctx, cancel := context.WithCancel(ctx)
-				created, err := createProcs(ctx, testutil.TrueCreateOpts(), client, 10)
+				opts := testutil.TrueCreateOpts()
+				modify(opts)
+				created, err := createProcs(ctx, opts, client, 10)
 				require.NoError(t, err)
 				assert.Len(t, created, 10)
 				cancel()
@@ -88,7 +96,9 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 		{
 			Name: "LongRunningOperationsAreListedAsRunning",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
-				procs, err := createProcs(ctx, testutil.SleepCreateOpts(20), client, 10)
+				opts := testutil.SleepCreateOpts(20)
+				modify(opts)
+				procs, err := createProcs(ctx, opts, client, 10)
 				require.NoError(t, err)
 				assert.Len(t, procs, 10)
 
@@ -108,7 +118,9 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 		{
 			Name: "ListReturnsOneSuccessfulCommand",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
-				proc, err := client.CreateProcess(ctx, testutil.TrueCreateOpts())
+				opts := testutil.TrueCreateOpts()
+				modify(opts)
+				proc, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
 
 				_, err = proc.Wait(ctx)
@@ -133,7 +145,9 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 		{
 			Name: "GetMethodReturnsMatchingDoc",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
-				proc, err := client.CreateProcess(ctx, testutil.TrueCreateOpts())
+				opts := testutil.TrueCreateOpts()
+				modify(opts)
+				proc, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
 
 				ret, err := client.Get(ctx, proc.ID())
@@ -152,7 +166,9 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 		{
 			Name: "GroupErrorsForCanceledContexts",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
-				_, err := client.CreateProcess(ctx, testutil.TrueCreateOpts())
+				opts := testutil.TrueCreateOpts()
+				modify(opts)
+				_, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
 
 				cctx, cancel := context.WithCancel(ctx)
@@ -166,7 +182,10 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 		{
 			Name: "GroupPropagatesMatching",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
-				proc, err := client.CreateProcess(ctx, testutil.TrueCreateOpts())
+				opts := testutil.TrueCreateOpts()
+				modify(opts)
+
+				proc, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
 
 				proc.Tag("foo")
@@ -189,8 +208,10 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 				if runtime.GOOS == "windows" {
 					t.Skip("the sleep tests don't block correctly on windows")
 				}
+				opts := testutil.SleepCreateOpts(100)
+				modify(opts)
 
-				_, err := createProcs(ctx, testutil.SleepCreateOpts(100), client, 10)
+				_, err := createProcs(ctx, opts, client, 10)
 				require.NoError(t, err)
 				assert.NoError(t, client.Close(ctx))
 			},
@@ -198,7 +219,10 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 		{
 			Name: "CloseErrorsWithCanceledContext",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
-				_, err := createProcs(ctx, testutil.SleepCreateOpts(100), client, 10)
+				opts := testutil.SleepCreateOpts(100)
+				modify(opts)
+
+				_, err := createProcs(ctx, opts, client, 10)
 				require.NoError(t, err)
 
 				cctx, cancel := context.WithCancel(ctx)
@@ -225,7 +249,10 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 		{
 			Name: "WaitingOnNonExistentProcessErrors",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
-				proc, err := client.CreateProcess(ctx, testutil.TrueCreateOpts())
+				opts := testutil.TrueCreateOpts()
+				modify(opts)
+
+				proc, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
 
 				_, err = proc.Wait(ctx)
@@ -244,6 +271,7 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 			Name: "ClearCausesDeletionOfProcesses",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
 				opts := testutil.TrueCreateOpts()
+				modify(opts)
 				proc, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
 				sameProc, err := client.Get(ctx, proc.ID())
@@ -261,6 +289,7 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 			Name: "ClearIsANoopForActiveProcesses",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
 				opts := testutil.SleepCreateOpts(20)
+				modify(opts)
 				proc, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
 				client.Clear(ctx)
@@ -274,10 +303,12 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 			Name: "ClearSelectivelyDeletesOnlyDeadProcesses",
 			Case: func(ctx context.Context, t *testing.T, client Manager) {
 				trueOpts := testutil.TrueCreateOpts()
+				modify(trueOpts)
 				lsProc, err := client.CreateProcess(ctx, trueOpts)
 				require.NoError(t, err)
 
 				sleepOpts := testutil.SleepCreateOpts(20)
+				modify(sleepOpts)
 				sleepProc, err := client.CreateProcess(ctx, sleepOpts)
 				require.NoError(t, err)
 
@@ -296,381 +327,376 @@ func AddBasicClientTests(tests ...ClientTestCase) []ClientTestCase {
 				require.NoError(t, jasper.Terminate(ctx, sleepProc)) // Clean up
 			},
 		},
-	}, tests...)
-}
-
-type ProcessTestCase struct {
-	Name       string
-	Case       func(context.Context, *testing.T, *options.Create, jasper.ProcessConstructor)
-	ShouldSkip bool
-}
-
-func AddBasicProcessTests(tests ...ProcessTestCase) []ProcessTestCase {
-	return append([]ProcessTestCase{
-		{
-			Name: "WithPopulatedArgsCommandCreationPasses",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				assert.NotZero(t, opts.Args)
-				proc, err := makep(ctx, opts)
+		ClientTestCase{
+			Name: "RegisterIsDisabled",
+			Case: func(ctx context.Context, t *testing.T, client Manager) {
+				err := client.Register(ctx, nil)
+				require.Error(t, err)
+				assert.Contains(t, err.Error(), "cannot register")
+			},
+		},
+		ClientTestCase{
+			Name: "CreateProcessReturnsCorrectExample",
+			Case: func(ctx context.Context, t *testing.T, client Manager) {
+				opts := testutil.TrueCreateOpts()
+				modify(opts)
+				proc, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
 				assert.NotNil(t, proc)
-			},
-		},
-		{
-			Name: "ErrorToCreateWithInvalidArgs",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				opts.Args = []string{}
-				proc, err := makep(ctx, opts)
-				require.Error(t, err)
-				assert.Nil(t, proc)
-			},
-		},
-		{
-			Name: "WithCanceledContextProcessCreationFails",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				pctx, pcancel := context.WithCancel(ctx)
-				pcancel()
-				proc, err := makep(pctx, opts)
-				require.Error(t, err)
-				assert.Nil(t, proc)
-			},
-		},
-		{
-			Name: "CanceledContextTimesOutEarly",
-			Case: func(ctx context.Context, t *testing.T, _ *options.Create, makep jasper.ProcessConstructor) {
-				pctx, pcancel := context.WithTimeout(ctx, 5*time.Second)
-				defer pcancel()
-				startAt := time.Now()
-				opts := testutil.SleepCreateOpts(20)
-				proc, err := makep(pctx, opts)
-				require.NoError(t, err)
-				require.NotNil(t, proc)
+				assert.NotZero(t, proc.ID())
 
-				time.Sleep(5 * time.Millisecond) // let time pass...
-				assert.False(t, proc.Info(ctx).Successful)
-				assert.True(t, time.Since(startAt) < 20*time.Second)
-			},
-		},
-		{
-			Name: "ProcessLacksTagsByDefault",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				tags := proc.GetTags()
-				assert.Empty(t, tags)
-			},
-		},
-		{
-			Name: "ProcessTagsPersist",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				opts.Tags = []string{"foo"}
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				tags := proc.GetTags()
-				assert.Contains(t, tags, "foo")
-			},
-		},
-		{
-			Name: "InfoHasMatchingID",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				_, err = proc.Wait(ctx)
-				require.NoError(t, err)
-				assert.Equal(t, proc.ID(), proc.Info(ctx).ID)
-			},
-		},
-		{
-			Name: "ResetTags",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				proc.Tag("foo")
-				assert.Contains(t, proc.GetTags(), "foo")
-				proc.ResetTags()
-				assert.Len(t, proc.GetTags(), 0)
-			},
-		},
-		{
-			Name: "TagsAreSetLike",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-
-				for i := 0; i < 50; i++ {
-					proc.Tag("foo")
-				}
-
-				assert.Len(t, proc.GetTags(), 1)
-				proc.Tag("bar")
-				assert.Len(t, proc.GetTags(), 2)
-			},
-		},
-		{
-			Name: "CompleteIsTrueAfterWait",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				time.Sleep(10 * time.Millisecond) // give the process time to start background machinery
-				_, err = proc.Wait(ctx)
+				fetched, err := client.Get(ctx, proc.ID())
 				assert.NoError(t, err)
-				assert.True(t, proc.Complete(ctx))
+				assert.NotNil(t, fetched)
+				assert.Equal(t, proc.ID(), fetched.ID())
 			},
 		},
-		{
-			Name: "WaitReturnsWithCanceledContext",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				opts.Args = []string{"sleep", "10"}
-				pctx, pcancel := context.WithCancel(ctx)
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				assert.True(t, proc.Running(ctx))
-				assert.NoError(t, err)
-				pcancel()
-				_, err = proc.Wait(pctx)
-				assert.Error(t, err)
-			},
-		},
-		{
-			Name: "RegisterTriggerErrorsForNil",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				assert.Error(t, proc.RegisterTrigger(ctx, nil))
-			},
-		},
-		{
-			Name: "RegisterSignalTriggerIDErrorsForExitedProcess",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				_, err = proc.Wait(ctx)
-				assert.NoError(t, err)
-				assert.Error(t, proc.RegisterSignalTriggerID(ctx, jasper.CleanTerminationSignalTrigger))
-			},
-		},
-		{
-			Name: "RegisterSignalTriggerIDFailsWithInvalidTriggerID",
-			Case: func(ctx context.Context, t *testing.T, _ *options.Create, makep jasper.ProcessConstructor) {
-				opts := testutil.SleepCreateOpts(3)
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				assert.Error(t, proc.RegisterSignalTriggerID(ctx, jasper.SignalTriggerID(-1)))
-			},
-		},
-		{
-			Name: "RegisterSignalTriggerIDPassesWithValidTriggerID",
-			Case: func(ctx context.Context, t *testing.T, _ *options.Create, makep jasper.ProcessConstructor) {
-				opts := testutil.SleepCreateOpts(3)
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				assert.NoError(t, proc.RegisterSignalTriggerID(ctx, jasper.CleanTerminationSignalTrigger))
-			},
-		},
-		{
-			Name: "WaitOnRespawnedProcessDoesNotError",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				require.NotNil(t, proc)
-				_, err = proc.Wait(ctx)
-				require.NoError(t, err)
-
-				newProc, err := proc.Respawn(ctx)
-				require.NoError(t, err)
-				_, err = newProc.Wait(ctx)
-				assert.NoError(t, err)
-			},
-		},
-		{
-			Name: "RespawnedProcessGivesSameResult",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				require.NotNil(t, proc)
-
-				_, err = proc.Wait(ctx)
-				require.NoError(t, err)
-				procExitCode := proc.Info(ctx).ExitCode
-
-				newProc, err := proc.Respawn(ctx)
-				require.NoError(t, err)
-				_, err = newProc.Wait(ctx)
-				require.NoError(t, err)
-				assert.Equal(t, procExitCode, newProc.Info(ctx).ExitCode)
-			},
-		},
-		{
-			Name: "RespawningFinishedProcessIsOK",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				require.NotNil(t, proc)
-				_, err = proc.Wait(ctx)
-				require.NoError(t, err)
-
-				newProc, err := proc.Respawn(ctx)
-				assert.NoError(t, err)
-				_, err = newProc.Wait(ctx)
-				require.NoError(t, err)
-				assert.True(t, newProc.Info(ctx).Successful)
-			},
-		},
-		{
-			Name: "RespawningRunningProcessIsOK",
-			Case: func(ctx context.Context, t *testing.T, _ *options.Create, makep jasper.ProcessConstructor) {
-				opts := testutil.SleepCreateOpts(2)
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				require.NotNil(t, proc)
-
-				newProc, err := proc.Respawn(ctx)
-				assert.NoError(t, err)
-				_, err = newProc.Wait(ctx)
-				require.NoError(t, err)
-				assert.True(t, newProc.Info(ctx).Successful)
-			},
-		},
-		{
-			Name: "RespawnShowsConsistentStateValues",
-			Case: func(ctx context.Context, t *testing.T, _ *options.Create, makep jasper.ProcessConstructor) {
-				opts := testutil.SleepCreateOpts(3)
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				require.NotNil(t, proc)
-				_, err = proc.Wait(ctx)
-				require.NoError(t, err)
-
-				newProc, err := proc.Respawn(ctx)
-				require.NoError(t, err)
-				assert.True(t, newProc.Running(ctx))
-				_, err = newProc.Wait(ctx)
-				require.NoError(t, err)
-				assert.True(t, proc.Complete(ctx))
-			},
-		},
-		{
-			Name: "WaitGivesSuccessfulExitCode",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, testutil.TrueCreateOpts())
-				require.NoError(t, err)
-				require.NotNil(t, proc)
-				exitCode, err := proc.Wait(ctx)
-				assert.NoError(t, err)
-				assert.Equal(t, 0, exitCode)
-			},
-		},
-		{
-			Name: "WaitGivesFailureExitCode",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, testutil.FalseCreateOpts())
-				require.NoError(t, err)
-				require.NotNil(t, proc)
-				exitCode, err := proc.Wait(ctx)
-				require.Error(t, err)
-				assert.Equal(t, 1, exitCode)
-			},
-		},
-		{
-			Name: "WaitGivesProperExitCodeOnSignalDeath",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, testutil.SleepCreateOpts(100))
-				require.NoError(t, err)
-				require.NotNil(t, proc)
-				sig := syscall.SIGTERM
-				require.NoError(t, proc.Signal(ctx, sig))
-				exitCode, err := proc.Wait(ctx)
-				require.Error(t, err)
-				if runtime.GOOS == "windows" {
-					assert.Equal(t, 1, exitCode)
-				} else {
-					assert.Equal(t, int(sig), exitCode)
-				}
-			},
-		},
-		{
-			Name: "WaitGivesNegativeOneOnAlternativeError",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				cctx, cancel := context.WithCancel(ctx)
-				proc, err := makep(ctx, testutil.SleepCreateOpts(100))
-				require.NoError(t, err)
-				require.NotNil(t, proc)
-
-				var exitCode int
-				waitFinished := make(chan bool)
-				go func() {
-					exitCode, err = proc.Wait(cctx)
-					waitFinished <- true
-				}()
-				cancel()
-				select {
-				case <-waitFinished:
-					require.Error(t, err)
-					assert.Equal(t, -1, exitCode)
-				case <-ctx.Done():
-					assert.Fail(t, "call to Wait() took too long to finish")
-				}
-				require.NoError(t, jasper.Terminate(ctx, proc)) // Clean up.
-
-			},
-		},
-		{
-			Name: "InfoHasTimeoutWhenProcessTimesOut",
-			Case: func(ctx context.Context, t *testing.T, _ *options.Create, makep jasper.ProcessConstructor) {
+		ClientTestCase{
+			Name: "WaitOnSigKilledProcessReturnsProperExitCode",
+			Case: func(ctx context.Context, t *testing.T, client Manager) {
 				opts := testutil.SleepCreateOpts(100)
-				opts.Timeout = time.Second
-				opts.TimeoutSecs = 1
-				proc, err := makep(ctx, opts)
+				modify(opts)
+				proc, err := client.CreateProcess(ctx, opts)
 				require.NoError(t, err)
+				require.NotNil(t, proc)
+				require.NotZero(t, proc.ID())
+
+				require.NoError(t, proc.Signal(ctx, syscall.SIGKILL))
 
 				exitCode, err := proc.Wait(ctx)
-				assert.Error(t, err)
+				require.Error(t, err)
 				if runtime.GOOS == "windows" {
 					assert.Equal(t, 1, exitCode)
 				} else {
-					assert.Equal(t, int(syscall.SIGKILL), exitCode)
+					assert.Equal(t, 9, exitCode)
 				}
-				info := proc.Info(ctx)
-				assert.True(t, info.Timeout)
 			},
 		},
-		{
-			Name: "CallingSignalOnDeadProcessDoesError",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
 
-				_, err = proc.Wait(ctx)
-				assert.NoError(t, err)
-
-				err = proc.Signal(ctx, syscall.SIGTERM)
-				require.Error(t, err)
-				assert.True(t, strings.Contains(err.Error(), "cannot signal a process that has terminated"))
-			},
-		},
-		{
-			Name: "CompleteAlwaysReturnsTrueWhenProcessIsComplete",
-			Case: func(ctx context.Context, t *testing.T, opts *options.Create, makep jasper.ProcessConstructor) {
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-
-				_, err = proc.Wait(ctx)
-				assert.NoError(t, err)
-
-				assert.True(t, proc.Complete(ctx))
-			},
-		},
-		{
-			Name: "RegisterSignalTriggerFails",
-			Case: func(ctx context.Context, t *testing.T, _ *options.Create, makep jasper.ProcessConstructor) {
-				opts := testutil.SleepCreateOpts(3)
-				proc, err := makep(ctx, opts)
-				require.NoError(t, err)
-				assert.Error(t, proc.RegisterSignalTrigger(ctx, func(_ jasper.ProcessInfo, _ syscall.Signal) bool {
-					return false
-				}))
+		ClientTestCase{
+			Name: "WriteFileFailsWithInvalidPath",
+			Case: func(ctx context.Context, t *testing.T, client Manager) {
+				opts := options.WriteFile{Content: []byte("foo")}
+				assert.Error(t, client.WriteFile(ctx, opts))
 			},
 		},
 	}, tests...)
+}
 
+func TestManager(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	for _, factory := range []struct {
+		Name        string
+		Constructor func(context.Context, *testing.T) Manager
+	}{
+		{
+			Name: "MDB",
+			Constructor: func(ctx context.Context, t *testing.T) Manager {
+				mngr, err := jasper.NewSynchronizedManager(false)
+				require.NoError(t, err)
+
+				client, err := makeTestMDBServiceAndClient(ctx, mngr)
+				require.NoError(t, err)
+				return client
+			},
+		},
+		{
+			Name: "RPC/TLS",
+			Constructor: func(ctx context.Context, t *testing.T) Manager {
+				mngr, err := jasper.NewSynchronizedManager(false)
+				require.NoError(t, err)
+
+				client, err := makeTLSRPCServiceAndClient(ctx, mngr)
+				require.NoError(t, err)
+				return client
+			},
+		},
+		{
+			Name: "RPC/Insecure",
+			Constructor: func(ctx context.Context, t *testing.T) Manager {
+				assert.NotPanics(t, func() {
+					newRPCClient(nil)
+				})
+
+				mngr, err := jasper.NewSynchronizedManager(false)
+				require.NoError(t, err)
+
+				client, err := makeInsecureRPCServiceAndClient(ctx, mngr)
+				require.NoError(t, err)
+				return client
+			},
+		},
+	} {
+		t.Run(factory.Name, func(t *testing.T) {
+			for _, modify := range []struct {
+				Name    string
+				Options testutil.OptsModify
+			}{
+				{
+					Name: "Blocking",
+					Options: func(opts *options.Create) {
+						opts.Implementation = options.ProcessImplementationBlocking
+					},
+				},
+				{
+					Name: "Basic",
+					Options: func(opts *options.Create) {
+						opts.Implementation = options.ProcessImplementationBasic
+					},
+				},
+				{
+					Name:    "Default",
+					Options: func(opts *options.Create) {},
+				},
+			} {
+				t.Run(modify.Name, func(t *testing.T) {
+					for _, test := range AddBasicClientTests(modify.Options,
+						ClientTestCase{
+							Name:       "StandardInput",
+							ShouldSkip: factory.Name == "MDB",
+							Case: func(ctx context.Context, t *testing.T, client Manager) {
+								for subTestName, subTestCase := range map[string]func(ctx context.Context, t *testing.T, opts *options.Create, expectedOutput string, stdin []byte){
+									"ReaderIsIgnored": func(ctx context.Context, t *testing.T, opts *options.Create, expectedOutput string, stdin []byte) {
+										opts.StandardInput = bytes.NewBuffer(stdin)
+
+										proc, err := client.CreateProcess(ctx, opts)
+										require.NoError(t, err)
+
+										_, err = proc.Wait(ctx)
+										require.NoError(t, err)
+
+										logs, err := client.GetLogStream(ctx, proc.ID(), 1)
+										require.NoError(t, err)
+										assert.Empty(t, logs.Logs)
+									},
+									"BytesSetsStandardInput": func(ctx context.Context, t *testing.T, opts *options.Create, expectedOutput string, stdin []byte) {
+										opts.StandardInputBytes = stdin
+
+										proc, err := client.CreateProcess(ctx, opts)
+										require.NoError(t, err)
+
+										_, err = proc.Wait(ctx)
+										require.NoError(t, err)
+
+										logs, err := client.GetLogStream(ctx, proc.ID(), 1)
+										require.NoError(t, err)
+
+										require.Len(t, logs.Logs, 1)
+										assert.Equal(t, expectedOutput, strings.TrimSpace(logs.Logs[0]))
+									},
+									"BytesCopiedByRespawnedProcess": func(ctx context.Context, t *testing.T, opts *options.Create, expectedOutput string, stdin []byte) {
+										opts.StandardInputBytes = stdin
+
+										proc, err := client.CreateProcess(ctx, opts)
+										require.NoError(t, err)
+
+										_, err = proc.Wait(ctx)
+										require.NoError(t, err)
+
+										logs, err := client.GetLogStream(ctx, proc.ID(), 1)
+										require.NoError(t, err)
+
+										require.Len(t, logs.Logs, 1)
+										assert.Equal(t, expectedOutput, strings.TrimSpace(logs.Logs[0]))
+
+										newProc, err := proc.Respawn(ctx)
+										require.NoError(t, err)
+
+										_, err = newProc.Wait(ctx)
+										require.NoError(t, err)
+
+										logs, err = client.GetLogStream(ctx, newProc.ID(), 1)
+										require.NoError(t, err)
+
+										require.Len(t, logs.Logs, 1)
+										assert.Equal(t, expectedOutput, strings.TrimSpace(logs.Logs[0]))
+									},
+								} {
+									t.Run(subTestName, func(t *testing.T) {
+										opts := &options.Create{
+											Args: []string{"bash", "-s"},
+											Output: options.Output{
+												Loggers: []options.Logger{jasper.NewInMemoryLogger(1)},
+											},
+										}
+										modify.Options(opts)
+
+										expectedOutput := "foobar"
+										stdin := []byte("echo " + expectedOutput)
+										subTestCase(ctx, t, opts, expectedOutput, stdin)
+									})
+								}
+							},
+						},
+						ClientTestCase{
+							Name:       "WriteFileSucceeds",
+							ShouldSkip: factory.Name == "MDB",
+							Case: func(ctx context.Context, t *testing.T, client Manager) {
+								tmpFile, err := ioutil.TempFile(buildDir(t), filepath.Base(t.Name()))
+								require.NoError(t, err)
+								defer func() {
+									assert.NoError(t, tmpFile.Close())
+									assert.NoError(t, os.RemoveAll(tmpFile.Name()))
+								}()
+
+								opts := options.WriteFile{Path: tmpFile.Name(), Content: []byte("foo")}
+								require.NoError(t, client.WriteFile(ctx, opts))
+
+								content, err := ioutil.ReadFile(tmpFile.Name())
+								require.NoError(t, err)
+
+								assert.Equal(t, opts.Content, content)
+							},
+						},
+						ClientTestCase{
+							Name:       "WriteFileAcceptsContentFromReader",
+							ShouldSkip: factory.Name == "MDB",
+							Case: func(ctx context.Context, t *testing.T, client Manager) {
+								tmpFile, err := ioutil.TempFile(buildDir(t), filepath.Base(t.Name()))
+								require.NoError(t, err)
+								defer func() {
+									assert.NoError(t, tmpFile.Close())
+									assert.NoError(t, os.RemoveAll(tmpFile.Name()))
+								}()
+
+								buf := []byte("foo")
+								opts := options.WriteFile{Path: tmpFile.Name(), Reader: bytes.NewBuffer(buf)}
+								require.NoError(t, client.WriteFile(ctx, opts))
+
+								content, err := ioutil.ReadFile(tmpFile.Name())
+								require.NoError(t, err)
+
+								assert.Equal(t, buf, content)
+							},
+						},
+						ClientTestCase{
+							Name:       "WriteFileSucceedsWithLargeContent",
+							ShouldSkip: factory.Name == "MDB",
+							Case: func(ctx context.Context, t *testing.T, client Manager) {
+								tmpFile, err := ioutil.TempFile(buildDir(t), filepath.Base(t.Name()))
+								require.NoError(t, err)
+								defer func() {
+									assert.NoError(t, tmpFile.Close())
+									assert.NoError(t, os.RemoveAll(tmpFile.Name()))
+								}()
+
+								const mb = 1024 * 1024
+								opts := options.WriteFile{Path: tmpFile.Name(), Content: bytes.Repeat([]byte("foo"), mb)}
+								require.NoError(t, client.WriteFile(ctx, opts))
+
+								content, err := ioutil.ReadFile(tmpFile.Name())
+								require.NoError(t, err)
+
+								assert.Equal(t, opts.Content, content)
+							},
+						},
+						ClientTestCase{
+							Name:       "WriteFileSucceedsWithNoContent",
+							ShouldSkip: factory.Name == "MDB",
+							Case: func(ctx context.Context, t *testing.T, client Manager) {
+								path := filepath.Join(buildDir(t), filepath.Base(t.Name()))
+								require.NoError(t, os.RemoveAll(path))
+								defer func() {
+									assert.NoError(t, os.RemoveAll(path))
+								}()
+
+								opts := options.WriteFile{Path: path}
+								require.NoError(t, client.WriteFile(ctx, opts))
+
+								stat, err := os.Stat(path)
+								require.NoError(t, err)
+
+								assert.Zero(t, stat.Size())
+							},
+						},
+
+						ClientTestCase{
+							Name:       "GetLogStreamFromNonexistentProcessFails",
+							ShouldSkip: factory.Name == "MDB",
+							Case: func(ctx context.Context, t *testing.T, client Manager) {
+								stream, err := client.GetLogStream(ctx, "foo", 1)
+								assert.Error(t, err)
+								assert.Zero(t, stream)
+							},
+						},
+						ClientTestCase{
+							Name:       "GetLogStreamFailsWithoutInMemoryLogger",
+							ShouldSkip: factory.Name == "MDB",
+							Case: func(ctx context.Context, t *testing.T, client Manager) {
+								opts := &options.Create{Args: []string{"echo", "foo"}}
+								modify.Options(opts)
+								proc, err := client.CreateProcess(ctx, opts)
+								require.NoError(t, err)
+								require.NotNil(t, proc)
+
+								_, err = proc.Wait(ctx)
+								require.NoError(t, err)
+
+								stream, err := client.GetLogStream(ctx, proc.ID(), 1)
+								assert.Error(t, err)
+								assert.Zero(t, stream)
+							},
+						},
+						ClientTestCase{
+							Name:       "WithInMemoryLogger",
+							ShouldSkip: factory.Name == "MDB",
+							Case: func(ctx context.Context, t *testing.T, client Manager) {
+								output := "foo"
+								opts := &options.Create{
+									Args: []string{"echo", output},
+									Output: options.Output{
+										Loggers: []options.Logger{
+											{
+												Type:    options.LogInMemory,
+												Options: options.Log{InMemoryCap: 100, Format: options.LogFormatPlain},
+											},
+										},
+									},
+								}
+								modify.Options(opts)
+
+								for testName, testCase := range map[string]func(ctx context.Context, t *testing.T, proc jasper.Process){
+									"GetLogStreamFailsForInvalidCount": func(ctx context.Context, t *testing.T, proc jasper.Process) {
+										stream, err := client.GetLogStream(ctx, proc.ID(), -1)
+										assert.Error(t, err)
+										assert.Zero(t, stream)
+									},
+									"GetLogStreamReturnsOutputOnSuccess": func(ctx context.Context, t *testing.T, proc jasper.Process) {
+										logs := []string{}
+										for stream, err := client.GetLogStream(ctx, proc.ID(), 1); !stream.Done; stream, err = client.GetLogStream(ctx, proc.ID(), 1) {
+											require.NoError(t, err)
+											require.NotEmpty(t, stream.Logs)
+											logs = append(logs, stream.Logs...)
+										}
+										assert.Contains(t, logs, output)
+									},
+								} {
+									t.Run(testName, func(t *testing.T) {
+										proc, err := client.CreateProcess(ctx, opts)
+										require.NoError(t, err)
+										require.NotNil(t, proc)
+
+										_, err = proc.Wait(ctx)
+										require.NoError(t, err)
+										testCase(ctx, t, proc)
+									})
+								}
+							},
+						},
+					) {
+						if test.ShouldSkip {
+							continue
+						}
+						t.Run(test.Name, func(t *testing.T) {
+							tctx, cancel := context.WithTimeout(ctx, testutil.RPCTestTimeout)
+							defer cancel()
+							test.Case(tctx, t, factory.Constructor(tctx, t))
+						})
+					}
+				})
+			}
+		})
+	}
 }
